+++ Gemfile
@@ -6,5 +6,6 @@ group :development do
   gem 'dotenv'
   gem 'tty-spinner'
   gem 'os'
+  gem 'colorize'
   gem 'pry'
 end
